requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.13.3
billiard==4.2.2
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.3
cryptography==44.0.2
cycler==0.12.1
daphne==4.1.2
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.9.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
fonttools==4.59.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
joblib==1.5.2
kiwisolver==1.4.9
kombu==5.5.4
matplotlib==3.10.6
msgpack==1.1.0
nltk==3.9.2
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
patsy==1.0.2
pillow==11.2.1
prompt_toolkit==3.0.52
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
redis==5.2.1
regex==2025.11.3
requests==2.32.3
scikit-learn==1.7.2
scipy==1.16.2
service-identity==24.2.0
setuptools==79.0.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
statsmodels==0.14.5
textblob==0.19.0
threadpoolctl==3.6.0
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.14
zope.interface==7.2
